requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.2
fake-bpy-module==20240721
idna==3.10
numpy==2.2.6
requests==2.32.4
urllib3==2.5.0
